Add command for finding new mint transactions
This will probably need iteration, but it works well on my current mint
data, so good enough for now! Happy to have this checkpoint done.

This library only works on python3. Getting the pickled data used in the
tests to work for both 2.7 and 3 was too much of a PITA. Maybe when we
move testing away from pickled data we can open this back up.